+++ oap-server/server-receiver-plugin/skywalking-sharing-server-plugin/src/main/java/org/apache/skywalking/oap/server/receiver/sharing/server/SharingServerModuleProvider.java
@@ -90,6 +90,7 @@ public class SharingServerModuleProvider extends ModuleProvider {
             setBootingParameter("oap.external.http.port", config.getRestPort());
 
             httpServer = new HTTPServer(httpServerConfig);
+            httpServer.setBlockingTaskName("receiver-http");
             httpServer.initialize();
 
             this.registerServiceImplementation(HTTPHandlerRegister.class, new HTTPHandlerRegisterImpl(httpServer));