+++ oap-server/server-receiver-plugin/zipkin-receiver-plugin/src/main/java/org/apache/skywalking/oap/server/receiver/zipkin/ZipkinReceiverProvider.java
@@ -88,6 +88,7 @@ public class ZipkinReceiverProvider extends ModuleProvider {
                                                                 .build();
 
             httpServer = new HTTPServer(httpServerConfig);
+            httpServer.setBlockingTaskName("zipkin-http");
             httpServer.initialize();
 
             httpServer.addHandler(